Add on-disk Cache

+++ src/image_creator/steps/base.py
@@ -1,33 +1,77 @@
-import shutil
+import pathlib
 from typing import Any, Dict
 
 from image_creator.constants import logger
 from image_creator.inputs import File
 from image_creator.steps import Step
 from image_creator.utils.download import download_file
-from image_creator.utils.misc import extract_xz_image, format_size, get_filesize
+from image_creator.utils.misc import (
+    copy_file,
+    extract_xz_image,
+    format_size,
+    get_filesize,
+)
 
 
 class DownloadImage(Step):
     name = "Fetching base image"
 
     def run(self, payload: Dict[str, Any]) -> int:
-        """whether system requirements are satisfied"""
-        base_file = payload["config"].base
+        # we need to extract this into the actual target
+        if payload["config"].base.getpath().suffix == ".xz":
+            return self.run_compressed(payload["config"].base, payload)
 
-        # no need to download
-        if base_file.is_local:
-            return self.run_local(base_file, payload)
+        return self.run_uncompressed(payload["config"].base, payload)
 
-        # we'll have to download it
-        return self.run_remote(base_file, payload)
+    def run_uncompressed(self, base_file: File, payload: Dict[str, Any]) -> int:
+        target = payload["options"].output_path
 
-    def run_remote(self, base_file: File, payload: Dict[str, Any]) -> int:
-        rem_path = base_file.getpath()
+        if base_file not in payload["cache"] and not base_file.is_local:
+            logger.start_task(f"Downloading {base_file.geturl()} into {target}…")
+            try:
+                download_file(base_file.geturl(), target)
+            except Exception as exc:
+                logger.fail_task(str(exc))
+                return 1
+            else:
+                logger.succeed_task(format_size(get_filesize(target)))
+                if payload["cache"].should_cache(base_file):
+                    logger.start_task("Adding Base Image to cache…")
+                    if payload["cache"].introduce(base_file, target):
+                        logger.succeed_task()
+                    else:
+                        logger.fail_task()
 
-        # we'll have to download it to build-dir then extract
-        if rem_path.suffix == ".xz":
-            xz_fpath = payload["options"].build_dir.joinpath(rem_path.name)
+            return 0
+
+        if base_file in payload["cache"]:
+            src_path = payload["cache"][base_file].fpath
+            message = f"Copying from cache… into {target}"
+        else:
+            src_path = base_file.getpath()
+            message = f"Copying {src_path} into {target}…"
+
+        logger.start_task(message)
+        try:
+            copy_file(src_path, target)
+        except Exception as exc:
+            logger.fail_task(str(exc))
+            return 1
+        else:
+            logger.succeed_task(format_size(get_filesize(target)))
+
+        if base_file in payload["cache"]:
+            payload["cache"][base_file] += 1
+
+        return 0
+
+    def run_compressed(self, base_file: File, payload: Dict[str, Any]) -> int:
+        target = payload["options"].output_path
+        xz_fpath = payload["options"].build_dir.joinpath(base_file.getpath().name)
+        remove_xz = False
+
+        # we need to download it first
+        if base_file not in payload["cache"] and not base_file.is_local:
             logger.start_task(f"Downloading {base_file.geturl()} into {xz_fpath}…")
             try:
                 download_file(base_file.geturl(), xz_fpath)
@@ -36,20 +80,43 @@ class DownloadImage(Step):
                 return 1
             else:
                 logger.succeed_task(format_size(get_filesize(xz_fpath)))
-
-            logger.start_task(
-                f"Extracting {xz_fpath} into {payload['options'].output_path}…"
+                remove_xz = True
+                if payload["cache"].should_cache(base_file):
+                    logger.start_task("Adding Base Image to cache…")
+                    if payload["cache"].introduce(base_file, xz_fpath):
+                        logger.succeed_task()
+                    else:
+                        logger.fail_task()
+        else:
+            xz_fpath = (
+                payload["cache"][base_file].fpath
+                if base_file in payload["cache"]
+                else base_file.getpath()
             )
-            try:
-                extract_xz_image(xz_fpath, payload["options"].output_path)
-            except Exception as exc:
-                logger.fail_task(str(exc))
-                return 1
-            else:
-                logger.succeed_task(
-                    format_size(get_filesize(payload["options"].output_path))
-                )
 
+        return self.extract(base_file, payload, xz_fpath, target, remove_xz)
+
+    def extract(
+        self,
+        base_file,
+        payload: Dict[str, Any],
+        xz_fpath: pathlib.Path,
+        target: pathlib.Path,
+        remove_xz: bool,
+    ) -> int:
+        logger.start_task(f"Extracting {xz_fpath}…")
+        try:
+            extract_xz_image(xz_fpath, target)
+        except Exception as exc:
+            logger.fail_task(str(exc))
+            return 1
+        else:
+            logger.succeed_task(format_size(get_filesize(target)))
+
+        if base_file in payload["cache"]:
+            payload["cache"][base_file] += 1
+
+        if remove_xz:
             logger.start_task(f"Removing {xz_fpath}…")
             try:
                 xz_fpath.unlink(missing_ok=True)
@@ -58,51 +125,4 @@ class DownloadImage(Step):
             else:
                 logger.succeed_task()
 
-        # download straight to final destination
-        else:
-            logger.start_task(
-                f"Downloading {base_file} into {payload['options'].output_path}…"
-            )
-            try:
-                download_file(base_file.geturl(), payload["options"].output_path)
-            except Exception as exc:
-                logger.fail_task(str(exc))
-                return 1
-            else:
-                logger.succeed_task(
-                    format_size(get_filesize(payload["options"].output_path))
-                )
-        return 0
-
-    def run_local(self, base_file: File, payload: Dict[str, Any]) -> int:
-        # we'll extract it to destination directly
-        if base_file.getpath().suffix == ".xz":
-            logger.start_task(
-                f"Extracting {base_file.getpath()} "
-                f"into {payload['options'].output_path}…"
-            )
-            try:
-                extract_xz_image(base_file.getpath(), payload["options"].output_path)
-            except Exception as exc:
-                logger.fail_task(str(exc))
-                return 1
-            else:
-                logger.succeed_task(
-                    format_size(get_filesize(payload["options"].output_path))
-                )
-        # we'll simply copy it to destination
-        else:
-            logger.start_task(
-                f"Copying {base_file.getpath()} "
-                f"into {payload['options'].output_path}…"
-            )
-            try:
-                shutil.copy2(base_file.getpath(), payload["options"].output_path)
-            except Exception as exc:
-                logger.fail_task(str(exc))
-                return 1
-            else:
-                logger.succeed_task(
-                    format_size(get_filesize(payload["options"].output_path))
-                )
         return 0
